ListView can be shown/hidden
The ListView pannel can be shown and hidden. This was implemented by adding
a new menu item, an action method to the document (as first responder) and
making the window controller added to the document on demand.
The list now also scrolls to the selected page and restores the scroll state
on reopening.