Nested zip file support
Added a dispatcher class method to CBPage that dispatches data+path page
creation. This is used from CPZipPage to create pages that aren't an image
(so can't use CBZipPage). That way, zip files inside zip files are supported
but require more memory/time due to no caching.
CBZipPage now tests on creation if the file name extension is supported by
NSImage, if it is not, creation fails and data page creation is attempted.
Also added CBDataPage, not used yet.