Added XAD Library support
XADMaster is a library that supports a ton of archive formats. It is licensed
under the LGPL, seems easy to use, so I hope it's perfect.
Added a new CBPage subclass that supports parsing with the XAD library, added
code that uses this class as fallback if ZipKit can't read it, and added
some meta data to the info.plist.