CBPage Refactoring
Moved image and NSDiscardableContent handling into CBParent, subclasses only
need to reimplement path getter and loadImage. This requires manually loading
the image with loadImage or using the beginContentAccess/endContentAccess
protocol. Note that image query methods like aspect/size and the like
give wrong data when the image is not loaded, so always begin access and
end it again.

Also fixed a memory leak with zip file loading and image loading.