Model<->View Page Update redesign
Changing the current page via the model now works. If the current page
changes during scrolling, the model is informed, if the model informs
of a page change, the position is changed if needed. The initial model
page is honored. This sometimes requires synchronous loading of pages
to ensure all are laid out.