Asynchronous Frame and Page Loading
Frames are now loaded in a separate thread and sent asynchronously to
the model in the main thread. By observing the frameCount the view
is informed and updates its pages with newly available images. If the
desired current page becomes available, the view jumps to it.
Decoding frames asynchronously showed that parts of XADM are not thread
safe, so access to each archive is synchronized.