CBComicLayout horizontal padding
Added horizontal padding support to the comic layout. The width of
the comic pages is now no longer 2.0 but 2.0 + padding. This required
some changes to the content layouting, which is now also explicitly
performed whenever the comic width might have changed.
Comic pages now have an explicit width instead of deriving it from
the anchor.