View Transform improvements
Clamping is now done directly in the zoomBy/moveBy modifiers instead
of a separate clamping method. This allows returning the effective
factor/offset, which is used in the zoomBy:withCenter method to keep
the mouse cursor pointed at the same thing. The modifiers are now
more clear in which coordinate system they expect their arguments to be.