Shift Pages
The "Shift Pages" functionality can change the even/odd page assignment.
It changes the current page from left to right side or vice versa, and
changes all other pages to match. This has no effect if the current page
is a double page. This does not change the layout order from left-to-right
or vice versa.