Current-Page detection
The current page is now detected by checking the distance of all
pages to the focus point instead of just checking vertical distance
to the anchor. This works properly with two side-by-side pages and
zoomed-in views.
New Geometry Helper functions for calculating distances between rects
and points are in CBGeometryHelpers.h.