Blocking Loading
Do not permit user interaction during loading. This seems like a
good idea, since the true order of pages is not yet known (thanks
to sorting at the end of the loading step). The user might change
the current page and then have it be changed again when the frames
are sorted.