CBPageLayer height limit
Limit the maximal height of a page by imposing a maximal aspect.
Pages that are taller are scaled to the size they would have if their
aspect would be in range. This is important for documents that contain
atypically long pages such as scanned book spines.
This limit is only in place for single pages in a double page layout.