Archive Extension Blacklist
Some file formats (in particular Tumbs.db) are supported by XAD, but
are not useful for this program, and cause significant negative
performance impact due to re-parsing the file repeatedly for each
entry, so they are blacklisted. This would be less of an issue with
full archive caching instead of only file archive caching.